+++ packages/components-angular/src/app/examples/index.ts
@@ -1,6 +1,8 @@
 import { AccordionExampleComponent } from './accordion-example.component';
 import { InlineNotificationExampleActionButtonComponent } from './inline-notification-example-action-button.component';
 import { InlineNotificationExampleEventsComponent } from './inline-notification-example-events.component';
+import { SegmentedControlExampleComponent } from './segmented-control-example.component';
+import { StepperHorizontalExampleComponent } from './stepper-horizontal-example.component';
 import { TableExampleAdvancedComponent } from './table-example-advanced.component';
 import { TableExampleBasicComponent } from './table-example-basic.component';
 import { TableExampleSortingComponent } from './table-example-sorting.component';
@@ -12,6 +14,8 @@ export const examples = [
   AccordionExampleComponent,
   InlineNotificationExampleActionButtonComponent,
   InlineNotificationExampleEventsComponent,
+  SegmentedControlExampleComponent,
+  StepperHorizontalExampleComponent,
   TableExampleAdvancedComponent,
   TableExampleBasicComponent,
   TableExampleSortingComponent,
@@ -23,6 +27,8 @@ export const examples = [
 export * from './accordion-example.component';
 export * from './inline-notification-example-events.component';
 export * from './inline-notification-example-action-button.component';
+export * from './segmented-control-example.component';
+export * from './stepper-horizontal-example.component';
 export * from './table-example-advanced.component';
 export * from './table-example-basic.component';
 export * from './table-example-sorting.component';