+++ packages/components/src/components/segmented-control/segmented-control/segmented-control-styles.spec.ts
@@ -1,0 +1,10 @@
+import { getComponentCss } from './segmented-control-styles';
+
+describe('getComponentCss()', () => {
+  it.each<Parameters<typeof getComponentCss>>([[20], [80], [230]])(
+    'should return correct css for maxWidth: %s',
+    (...args) => {
+      expect(getComponentCss(...args)).toMatchSnapshot();
+    }
+  );
+});