+++ packages/shared/scripts/generateCodeSamples.ts
@@ -42,6 +42,26 @@ const generateCodeSamples = (): void => {
       ],
     },
     {
+      component: 'p-segmented-control',
+      samples: [
+        [
+          componentsJsPath + '/segmented-control-example.html',
+          componentsAngularPath + '/segmented-control-example.component.ts',
+          componentsReactPath + '/SegmentedControlExample.tsx',
+        ],
+      ],
+    },
+    {
+      component: 'p-stepper-horizontal',
+      samples: [
+        [
+          componentsJsPath + '/stepper-horizontal-example.html',
+          componentsAngularPath + '/stepper-horizontal-example.component.ts',
+          componentsReactPath + '/StepperHorizontalExample.tsx',
+        ],
+      ],
+    },
+    {
       component: 'p-table',
       samples: [
         [